Refactored for clarity.
