+++ src/inwatch.ts
@@ -1,9 +1,15 @@
 import { Watch, WatchOptions } from "./watch"
+import { Wait, WaitOptions } from "./wait"
 
+export * from "./wait"
 export * from "./watch"
-export * from "./monitor"
+
+/** Create a high-level watch that emits `add`, `change`, and `remove` events for files. */
+export default function watch(path: string, options?: WatchOptions) {
+  return new Watch(path, options)
+}
 
 /** Spawn `inotifywait` in monitor mode to parse and emit events. */
-export default function watch(paths: string | string[], options?: WatchOptions) {
-  return new Watch(paths, options)
+export function wait(paths: string | string[], options?: WaitOptions) {
+  return new Wait(paths, options)
 }
